% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_mncolors.R
\name{scale_fill_mncol}
\alias{scale_fill_mncol}
\title{Color Palette MN Colors - Fill}
\usage{
scale_fill_mncol(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{The name of the palette; options are "main", "blue", and "green".}

\item{discrete}{Boolean condition to define whether relevant aesthetic is discrete or not. Default is false.}

\item{reverse}{Boolean condition to reverse order of colors or not. Default is false.}
}
\value{
Outputs to SQLite database all .csv files with an additional column of row numbers to uniquely identify records (sometimes redundantly).
}
\description{
This function uses interpolated MN colors to form a color palette for graphical purposes. The three main palettes are "main", "blue", and "green".
"Main" is composed of MN branded colors, using interpolated colors for gradiation, sorted in the arrangement of the color wheel (with a small deviation -- can you see it!?)
}
